move e2e utils to common location
