+++ packages/react-examples/src/react/FocusTrapZone/e2e/FocusTrapZone.NoTabbableItems.stories.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { FocusTrapZone, mergeStyles } from '@fluentui/react';
-import { useProps } from './shared';
+import type { IFocusTrapZoneProps } from '@fluentui/react';
+import { useProps } from '../../../e2e/utils';
 
 const rootClass = mergeStyles({
   button: { height: 30, width: 60 },
@@ -11,7 +12,7 @@ const rootClass = mergeStyles({
  * Tab and shift-tab when the FTZ contains 0 tabbable items
  */
 export const NoTabbableItems = () => {
-  const props = useProps();
+  const props = useProps<IFocusTrapZoneProps>();
 
   return (
     // don't render until props have been set
